Fixed wrong IDs in FormatType Added WebSeries movie type Fixed decimal places in ratings being dropped Fixed actors for a movie appearing in the same order as on IMDB. Fixed Java 1.5 compatibility by removing 1.6 specific code
DATABASE/FILE FORMAT CHANGES
Databases created will now have a new field CHARACTERNUMBER in table MOVIEACTOR
MovieType ID's are now different from before

+++ src/com/googlecode/jmoviedb/model/movietype/AbstractMovie.java
@@ -462,7 +462,7 @@ public abstract class AbstractMovie implements Cloneable {
 	}
 	
 	public void setRatingAsInt(int rating) {
-		this.rating = rating/10;
+		this.rating = (double)rating/10;
 	}
 
 	public int getRunTime() {
